requirements.txt
asgiref==3.4.1
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.0
dj-database-url==0.5.0
Django==4.0
django-extensions==3.1.5
django-filter==21.1
django-vectortiles==0.1.0
djangorestframework==3.13.1
djangorestframework-gis==0.17
djangorestframework-mvt==0.2.5
gunicorn==20.1.0
idna==3.3
ipython==7.30.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mercantile==1.2.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.24
psycopg2-binary==2.9.2
ptyprocess==0.7.0
Pygments==2.10.0
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
traitlets==5.1.1
uritemplate==4.1.1
urllib3==1.26.7
uWSGI==2.0.20
wcwidth==0.2.5

